requirements.txt
Django==4.2.7
torch==2.0.1
torchvision==0.15.2
transformers==4.35.2
Pillow==10.1.0
numpy==1.24.3
opencv-python==4.8.1.78
ultralytics==8.0.200
accelerate==0.24.1
jupyter==1.0.0
hf_xet==0.1.0